+++ resources/js/Components/Tokens/TokenDetails.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { Heading } from "@/Components/Heading";
-import { useActiveUser } from "@/Contexts/ActiveUserContext";
+import { useAuth } from "@/Contexts/AuthContext";
 import { assertUser } from "@/Utils/assertions";
 import { FormatFiat, FormatNumber } from "@/Utils/Currency";
 
@@ -10,7 +10,8 @@ interface Properties extends React.HTMLAttributes<HTMLDivElement> {
 
 export const TokenDetails = ({ token, ...properties }: Properties): JSX.Element => {
     const { t } = useTranslation();
-    const { user } = useActiveUser();
+    const { user } = useAuth();
+
     assertUser(user);
 
     const currency = user.attributes.currency;
